requirements.txt
requests
numpy
aiogram>=3.0
aiohttp
python-dotenv
